Add folders and move services
